Refactoring: rename podLogs to controllers/logCtrl
