+++ packages/graphql-server/src/ee/app.ts
@@ -41,7 +41,7 @@ import cors from '@koa/cors';
 import { JobLogCtrl } from './controllers/jobLogCtrl';
 import { PhJobCacheList } from './crdClient/phJobCacheList';
 import JobArtifactCleaner from './utils/jobArtifactCleaner';
-import { PodLogs } from '../utils/podLogs';
+import { PodLogs } from '../controllers/logCtrl';
 
 // cache
 import {
